Edited "do_input()" for solving "2kino"-Bug from issue 1
After a valid input into variable, std::cin will be emptied until next
"\n". Therefore you won't be able to use the template with strings,
from now on, as the input might be taken until "\n", and then the
input until the NEXT \n (which will be at next input) will be
discarded!